pyproject.toml
[project]
name = "ha-boneco"
version = "0.1.0"
description = "A Home Assistant integration to control Boneco devices"
readme = "README.md"
authors = [
    { name = "DeKaN" }
]
requires-python = ">=3.13"
dependencies = [
    "pyboneco==0.3.8",
]

[project.urls]
Repository = "https://github.com/DeKaN/ha-boneco.git"

[dependency-groups]
dev = [
    "homeassistant-stubs==2025.3.0",
    "pre-commit==4.2.0",
    "ruff==0.9.7",
]
